+++ src/main/kotlin/org/olafneumann/regex/generator/ui/HtmlHelper.kt
@@ -7,9 +7,12 @@ import org.w3c.dom.HTMLElement
 import kotlinx.browser.document
 import kotlinx.browser.window
 import org.olafneumann.regex.generator.RegexGeneratorException
+import org.olafneumann.regex.generator.js.JQuery
 import org.olafneumann.regex.generator.js.encodeURIComponent
 import org.olafneumann.regex.generator.regex.RecognizerMatch
+import org.olafneumann.regex.generator.ui.HtmlView.Companion.each
 import org.w3c.dom.url.URL
+import kotlin.math.max
 
 internal object HtmlHelper {
     internal inline fun <reified T : HTMLElement> getElementById(id: String): T {
@@ -18,6 +21,15 @@ internal object HtmlHelper {
         } catch (e: ClassCastException) {
             throw RegexGeneratorException("Unable to find element with id '${id}'.", e)
         }
+    }
+
+    internal fun getHeight(elements: JQuery): Int {
+        val previousCss = elements.attr("style")
+        elements.css("position:absolute;visibility:hidden;display:block !important;")
+        var maxHeight = 0
+        elements.each { jq -> maxHeight = max(maxHeight, jq.height()) }
+        elements.attr("style", previousCss ?: "")
+        return maxHeight
     }
 }
